reordering

+++ Models/Item.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel, ConfigDict
-from typing import Optional, List
+from typing import Optional, List, Any
 from datetime import datetime
 
 
@@ -7,7 +7,7 @@ class Item(BaseModel):
     model_config = ConfigDict(
         arbitrary_types_allowed=True
     )
-    price: Optional[float]
+    price: Optional[int]
     name: Optional[str]
     category: Optional[str]
     thumbnail_url: Optional[str]
